backend/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
python-multipart

websockets
